Move api logging type to APIServerLog module
