+++ hydra-node/src/Hydra/Logging/Messages.hs
@@ -10,11 +10,11 @@ module Hydra.Logging.Messages where
 
 import Hydra.Prelude
 
+import Hydra.API.APIServerLog (APIServerLog)
 import Hydra.Chain.Direct.Handlers (DirectChainLog)
 import Hydra.Network.Authenticate (AuthLog)
 import Hydra.Node (HydraNodeLog)
 import Hydra.Options (ParamMismatch, RunOptions)
-import Hydra.API.HTTPServer (APIServerLog)
 
 data HydraLog tx net
   = DirectChain {directChain :: DirectChainLog}
